Ref #5222 Store the uncommitted documents in a module global for MockSolrInterface
I think this is more realistic according to solr - my understanding is
that commit causes all documents submitted, regardless of the client, to
be committed.
Without this change the tests broke - as the reworking of Backend(Queue)
currently means that a different sunburnt.SolrInterface instance is used
for add() vs. commit().

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@25119 484df14f-75d2-4b3f-8ba0-490e05069a01